Fixes #48419 by cleaning the whitelist of broken links
https://savannah.gnu.org/bugs/index.php?48419

Updates some of the dojo.js links in the whitelist

As reported in https://savannah.gnu.org/bugs/index.php?48419, some of the
scripts listed in the LibreJS whitelist are no longer available at the listed
URLs.

Signed-off-by: zachwick <zach@zachwick.com>

Removes toolbarbutton-jplib as it can no longer be found anywhere

Signed-off-by: zachwick <zach@zachwick.com>

Removes link to the minified version of underscore.js v1.1.6

I cannot find a canonical/confirmable minified version of this version of
underscore.js.

Signed-off-by: zachwick <zach@zachwick.com>

Removes v3.0.2 of cryptojs from the whitelist

The Google Code version of this code is no longer available, and there exists no
canonical and verifiable replacement location of this project which contains
this version of crpytojs.

Signed-off-by: zachwick <zach@zachwick.com>

Removes v3.3.0 of yui-base from the whitelist

This version of yui-base can no longer be found at the listed URL. No canonical
verifiable URL for this version can be found to use as a replacement.

It is worth noting that the YUI project is no longer actively maintained, and
long term we may want to fully remove YUI from the whitelist.

Signed-off-by: zachwick <zach@zachwick.com>